feat: enhance forwarding logs and DND management
- Include 'CALL_FORWARDED' status in log queries
- Add DND permission request flow in settings
- Ensure scheduled SMS tasks are logged to the database
- Fix task ID type mismatch in WorkManager data

+++ app/src/main/java/com/example/ui/screens/ScheduleScreen.kt
@@ -202,7 +202,7 @@ fun ScheduleScreen(viewModel: MainViewModel, onNavigateToAdd: () -> Unit) {
                                         val delay = timeMillis - System.currentTimeMillis()
                                         if (delay > 0) {
                                             val data = Data.Builder()
-                                                .putLong("taskId", id)
+                                                .putInt("taskId", id.toInt())
                                                 .putString("type", type)
                                                 .putString("target", target)
                                                 .putString("message", msg)
